Fix integer overflow on freqs >2.14 GHz.
This changes all variables that store the tuner center frequency to use
int64_t to avoid overflowing 32 bit signed ints.

It fixes the following error:

$ ./rtl_power_fftw -f 2150M
Invalid frequency given to --freq: -2147483648.
Expecting a positive number, allowing the k,M,G multipliers. Exiting.

+++ src/acquisition.cxx
@@ -213,7 +213,7 @@ Acquisition::Acquisition(const Params& params_,
                          Rtlsdr& rtldev_,
                          Datastore& data_,
                          int actual_samplerate_,
-                         int freq_) :
+                         int64_t freq_) :
   params(params_), aux(aux_), rtldev(rtldev_), data(data_),
   actual_samplerate(actual_samplerate_), freq(freq_)
 { }
